Split THStorage into class and functions
