+++ aten/src/TH/TH.h
@@ -12,7 +12,7 @@
 #include "THLogAdd.h"
 #include "THRandom.h"
 #include "THSize.h"
-#include "THStorage.h"
+#include "THStorageFunctions.h"
 #include "THTensor.h"
 #include "THTensorApply.h"
 #include "THTensorDimApply.h"
