+++ aten/src/TH/THFile.cpp
@@ -1,5 +1,5 @@
 #include "THFile.h"
-#include "THStorage.hpp"
+#include "THStorageFunctions.hpp"
 #include "THFilePrivate.h"
 
 #define IMPLEMENT_THFILE_RW(TYPEC, TYPE)                          \
