+++ aten/src/TH/THMemoryFile.cpp
@@ -1,5 +1,5 @@
 #include "THMemoryFile.h"
-#include "THStorage.hpp"
+#include "THStorageFunctions.hpp"
 #include "THFilePrivate.h"
 #include "THDiskFile.h"
 #include "stdint.h"
