+++ aten/src/TH/THMemoryFile.h
@@ -2,7 +2,7 @@
 #define TH_MEMORY_FILE_INC
 
 #include "THFile.h"
-#include "THStorage.h"
+#include "THStorageFunctions.h"
 
 TH_API THFile *THMemoryFile_newWithStorage(THCharStorage *storage, const char *mode);
 TH_API THFile *THMemoryFile_new(const char *mode);
