+++ aten/src/TH/THTensor.h
@@ -1,7 +1,7 @@
 #ifndef TH_TENSOR_INC
 #define TH_TENSOR_INC
 
-#include "THStorage.h"
+#include "THStorageFunctions.h"
 #include "THTensorApply.h"
 
 #define THTensor_(NAME)   TH_CONCAT_4(TH,Real,Tensor_,NAME)
