+++ aten/src/THC/THCStorage.h
@@ -1,7 +1,7 @@
 #ifndef THC_STORAGE_INC
 #define THC_STORAGE_INC
 
-#include "THStorage.h"
+#include "THStorageFunctions.h"
 #include "THCGeneral.h"
 
 #define THCStorage_(NAME) TH_CONCAT_4(TH,CReal,Storage_,NAME)
